# openapi.json
{"openapi":"3.0.0","info":{"title":"Hotel API","version":"0.0.1"},"servers":[{"url":"http://localhost:8092/property-view","description":"Generated server url"}],"paths":{"/hotels":{"get":{"tags":["hotels"],"summary":"Get all hotels","operationId":"getAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HotelShortResponse"}}}}}}},"post":{"tags":["hotels"],"summary":"Create new hotel","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHotel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelShortResponse"}}}},"404":{"description":"Data is not valid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/hotels/{hotelId}/amenities":{"post":{"tags":["amenities"],"summary":"Add amenities for hotel","operationId":"update","parameters":[{"name":"hotelId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK"},"404":{"description":"Hotel not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/search":{"get":{"tags":["hotels"],"summary":"Search hotels","operationId":"search","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"brand","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"city","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"county","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"amenities","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HotelShortResponse"}}}}}}}},"/hotels/{id}":{"get":{"tags":["hotels"],"summary":"Get hotel by id","operationId":"find","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelResponse"}}}},"404":{"description":"Hotel not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/histogram/{parameter}":{"get":{"tags":["histograms"],"summary":"Get the number of hotels grouped by parameter","operationId":"groupBy","parameters":[{"name":"parameter","in":"path","required":true,"schema":{"type":"string","enum":["brand","city","county","amenities"]}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}}},"components":{"schemas":{"Address":{"type":"object","properties":{"houseNumber":{"type":"integer","format":"int32","minimum":1},"street":{"type":"string"},"city":{"type":"string"},"county":{"type":"string"},"postCode":{"type":"string"}},"required":["city","county","postCode","street"]},"ArrivalTime":{"type":"object","properties":{"checkIn":{"type":"string","example":"12:00"},"checkOut":{"type":"string","example":"12:00"}},"required":["checkIn"]},"Contacts":{"type":"object","properties":{"phone":{"type":"string"},"email":{"type":"string"}},"required":["email","phone"]},"CreateHotel":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"brand":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"},"contacts":{"$ref":"#/components/schemas/Contacts"},"arrivalTime":{"$ref":"#/components/schemas/ArrivalTime"}},"required":["address","arrivalTime","brand","contacts","name"]},"ErrorResponse":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"path":{"type":"string"}}},"HotelShortResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"}}},"HotelResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"brand":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"},"contacts":{"$ref":"#/components/schemas/Contacts"},"arrivalTime":{"$ref":"#/components/schemas/ArrivalTime"},"amenities":{"type":"array","items":{"type":"string"}}}}}}}